order_service/requirements.txt
alembic
python-dotenv
sqlalchemy
psycopg2
asyncpg
pydantic
pydantic_settings
fastapi
uvicorn